Lua: Migrate utility functions to separate files
